+++ src/csv/mod.rs
@@ -1,2 +1,3 @@
 pub mod csv_reader;
 pub mod csv_writer;
+pub mod reader_slice;