+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ---
-[1.0.1] - 2025-05-04
+[1.0.3] - 2025-05-05
+
+- Added CsvReader from Slice, where you dont depend on mmap.
+- Added mutable param to use inside worker.Util to avoid locking
+- Changed peek on iterator to not be mutable. Now get field, not need a mutable row.
+- Non mutable peek, allow to repair macros get_Str, get string, get_str_utf8
+- Downgrade edition to 2018, to add compatibility.
+
+---
+
+[1.0.2] - 2025-05-04
 
 - Removed string, cow, &str macros